+++ jest.config.ts
@@ -5,10 +5,6 @@ import {compilerOptions} from './tsconfig.json'
 
 export default async (): Promise<Config.InitialOptions> => {
   return {
-    globals: {
-      hb: true
-    },
-
     transform: {
       '^.+\\.ts?$': [
         'ts-jest',
@@ -31,7 +27,7 @@ export default async (): Promise<Config.InitialOptions> => {
     moduleNameMapper: pathsToModuleNameMapper(compilerOptions.paths),
     modulePaths: ['<rootDir>'],
     forceExit: true,
-    clearMocks: true,
+    restoreMocks: true,
     testTimeout: 15000
   }
 }